chore(connect): firmware utils rearrangement
